Removed old fibonacci implementation and renamed quick_fibonacci.py to fibonacci.py
